pyproject.toml
[project]
name = "extract-score"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10.6",
    "python-dateutil>=2.9.0.post0",
]
